Colore: transform DF to ExtPointRDD

+++ Colore/src/main/scala/ca/ExtPoint.scala
@@ -24,7 +24,7 @@ import com.vividsolutions.jts.geom.Envelope
 /*
  */
 
-class ExtPoint(point: Point, z: Float, dz: Float) extends Point(point.getCoordinateSequence, point.getFactory)
+class ExtPoint(point: Point, z: Double, dz: Double) extends Point(point.getCoordinateSequence, point.getFactory)
 {
     override def toString: String = super.toString + " z=" + z + " dz=" + dz
     def getPoint = this.point
